+++ src/math/mod.rs
@@ -16,6 +16,16 @@ mod libm_math {
     }
 
     #[inline(always)]
+    pub(crate) fn ceil(f: f32) -> f32 {
+        libm::ceilf(f)
+    }
+
+    #[inline(always)]
+    pub(crate) fn round(f: f32) -> f32 {
+        libm::roundf(f)
+    }
+
+    #[inline(always)]
     pub(crate) fn sin(f: f32) -> f32 {
         libm::sinf(f)
     }
@@ -56,6 +66,16 @@ mod std_math {
     #[inline(always)]
     pub(crate) fn floor(f: f32) -> f32 {
         f32::floor(f)
+    }
+
+    #[inline(always)]
+    pub(crate) fn ceil(f: f32) -> f32 {
+        f32::ceil(f)
+    }
+
+    #[inline(always)]
+    pub(crate) fn round(f: f32) -> f32 {
+        f32::round(f)
     }
 
     #[inline(always)]